requirements.txt
streamlit
ctransformers
faiss-cpu
langchain==0.0.331
pypdf
python-box
sentence-transformers